+++ db/migrate/20181018042401_create_user_ref_codes.rb
@@ -1,8 +1,8 @@
 class CreateUserRefCodes < ActiveRecord::Migration[5.1]
   def change
     create_table :user_ref_codes do |t|
-      t.integer :user_id
-      t.integer :ref_code_id
+      t.bigint :user_id
+      t.bigint :ref_code_id
 
       t.timestamps
     end
